Unique the device list.
The device list is no longer a list of objects; instead it's a list of device
names converted at the destination to the old format. This was the only way to
be able to use the default set otherwise the list of objects would compare
equivalent as not equal and allow duplicates. This closes #15.

As well, the PeerList now directly subscribes to the redux store. It checks
for updates to the devices list and consistently updates with new data. This
closes #76.