a few network component changes
the network component is now only declared in the select mode screen. Since this screen is at the top of the stack navigator, the network component is never unmounted (which in this case works to our advantage of only having to instantiate the network component once in our application (kind of like a singleton). However, this has the added side effect of leaving the PeerList also instantiated; I don't think this is a really huge deal though. I have added timestamps to all of our messages sent over google nearby, and logic to the network component that now ensures devices stay in sync in terms of values for heart rate, BP, etc.